pyproject.toml
[tool.poetry]
name = "Fire_Alarm_Systems_Back_End"
version = "0.1.0"
description = "Your project description"
authors = ["RonAsron <ronasron2546@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2"
djangorestframework = "^3.14"
gunicorn = "^23.0.0"
django-cors-headers = "^4.6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
